requirements.txt
dgl_cu102==0.5.2
line_profiler==4.0.1
numpy==1.21.6
pandas==1.3.5
PyYAML==6.0
rdflib==6.2.0
scikit_learn==1.0.2
scipy==1.7.3
torch==1.6.0
tqdm==4.64.1
